refactor: ApplicationCard의 함수들 libs폴더에 분리
